requirements.txt
torch
transformers
peft
tqdm
pillow
faiss-cpu  # or faiss-gpu if available